+++ backend/api/routes/public/routes_auth.py
@@ -5,16 +5,18 @@ from aiohttp import web
 from pydantic import ValidationError
 from sqlalchemy.exc import SQLAlchemyError
 
-from api.types import RegistrationPayload, ConfirmCodePayload
+from globalTypes.consts import Role
+
+from .types import RegistrationPayload, ConfirmCodePayload, LoginPayload
 from database import repositories
-from globalTypes import RegistrationType, ConfirmStatusCode
-from pkg import jwt, errors
-from pkg import rabbitmq
+from globalTypes import RegistrationType, ConfirmStatusCode, AuthStatus
+from pkg import jwt, errors, rabbitmq, password
 
 
 def setup_auth_routes(app: web.Application):
     app.router.add_post("/api/auth/register", handle_registration)
     app.router.add_post("/api/auth/confirm", handle_registration_confirm)
+    app.router.add_post("/api/auth/login", handle_login)
 
 
 async def handle_registration(request: web.Request):
@@ -24,7 +26,8 @@ async def handle_registration(request: web.Request):
     except ValidationError as err:
         return web.json_response(
             json.loads(
-                err.json(include_url=False, include_input=False, include_context=False)
+                err.json(include_url=False, include_input=False,
+                         include_context=False)
             ),
             status=HTTPStatus.BAD_REQUEST,
         )
@@ -33,18 +36,12 @@ async def handle_registration(request: web.Request):
         try:
             if payload.type == RegistrationType.CANDIDATE:
                 user, candidate = await repositories.create_candidate(session, payload)
-                rabbitmq.admin_create_new_candidate(request.app["mq"], user.json())
+                rabbitmq.admin_create_new_candidate(
+                    request.app["mq"], user.json())
                 rabbitmq.email_confirm_account(
                     request.app["mq"], user.json(), user.codes[0].json()
                 )
-                return web.json_response(
-                    {
-                        "token": jwt.create_jwt_token(user),
-                        "user": user.json(),
-                        "candidate": candidate.json(),
-                    },
-                    status=HTTPStatus.OK,
-                )
+                return web.json_response({"status": True}, status=HTTPStatus.OK)
 
             if payload.type == RegistrationType.COMPANY:
                 if not payload.companyName:
@@ -62,21 +59,17 @@ async def handle_registration(request: web.Request):
                     session, payload
                 )
                 rabbitmq.admin_create_new_company(
-                    request.app["mq"], company.json(), member.json(), user.json()
+                    request.app["mq"], company.json(
+                    ), member.json(), user.json()
                 )
                 rabbitmq.email_confirm_account(
                     request.app["mq"], user.json(), user.codes[0].json()
                 )
-                return web.json_response(
-                    {
-                        "token": jwt.create_jwt_token(user),
-                        "user": user.json(),
-                        "company": company.json(),
-                    },
-                    status=HTTPStatus.OK,
-                )
+                return web.json_response({"status": True}, status=HTTPStatus.OK)
 
         except SQLAlchemyError as e:
+            await session.rollback()
+
             if "duplicate key value violates unique" in e.args[0]:
                 # TODO: find out what field trigger this error.
                 # Because companyName also should be unique, but
@@ -93,6 +86,11 @@ async def handle_registration(request: web.Request):
                 {"errors": [str(e)]}, status=HTTPStatus.BAD_REQUEST
             )
 
+    return web.json_response(
+        {"errors": "something went wrong"}, status=HTTPStatus.BAD_REQUEST
+    )
+
+
 
 async def handle_registration_confirm(request: web.Request):
     data = await request.json()
@@ -101,7 +99,8 @@ async def handle_registration_confirm(request: web.Request):
     except ValidationError as err:
         return web.json_response(
             json.loads(
-                err.json(include_url=False, include_input=False, include_context=False)
+                err.json(include_url=False, include_input=False,
+                         include_context=False)
             ),
             status=HTTPStatus.BAD_REQUEST,
         )
@@ -116,12 +115,6 @@ async def handle_registration_confirm(request: web.Request):
 
             # TODO: is expired - return expired error and delete code.
 
-            if code.code != payload.code:
-                return web.json_response(
-                    [errors.create_form_error("code", "Code mismatch")],
-                    status=HTTPStatus.NOT_FOUND,
-                )
-
             # Call confirm_user function to confirm the user and update the status
             user = await repositories.confirm_user(session, code)
             if user is None:
@@ -131,8 +124,10 @@ async def handle_registration_confirm(request: web.Request):
 
             # Prepare payloads and publish success messages to RabbitMQ
             rabbitmq_payloads = user.json()
-            rabbitmq.email_confirm_account_success(request.app["mq"], rabbitmq_payloads)
-            rabbitmq.admin_confirm_account_success(request.app["mq"], rabbitmq_payloads)
+            rabbitmq.email_confirm_account_success(
+                request.app["mq"], rabbitmq_payloads)
+            rabbitmq.admin_confirm_account_success(
+                request.app["mq"], rabbitmq_payloads)
 
             return web.json_response({"status": "confirmed"}, status=HTTPStatus.OK)
 
@@ -141,4 +136,44 @@ async def handle_registration_confirm(request: web.Request):
             return web.json_response(
                 {"errors": [str(e)]}, status=HTTPStatus.BAD_REQUEST
             )
+
+
+async def handle_login(request: web.Request):
+    data = await request.json()
+    try:
+        payload = LoginPayload(**data)
+    except ValidationError as err:
+        return web.json_response(
+            json.loads(
+                err.json(include_url=False, include_input=False,
+                         include_context=False)
+            ),
+            status=HTTPStatus.BAD_REQUEST,
+        )
+
+    async with request.app["session"] as session:
+        user = await repositories.get_user(session, **{
+            "email": payload.email,
+            "active": True,
+            "status": AuthStatus.ACTIVE,
+        })
+
+        if user is None:
+            return web.json_response({"error": "not found"}, status=HTTPStatus.NOT_FOUND)
+
+        if not password.check_password(payload.password, user.password_hash):
+            return web.json_response({"error": "Wrong credentials"}, status=HTTPStatus.NOT_FOUND)
         
+        company = None
+        if user.role in [Role.COMPANY_ADMIN, Role.COMPANY_MANAGER]:
+            company = await repositories.get_company(session, id=user.manager.company_id)
+
+    resp = {
+        "token": jwt.create_jwt_token(user),
+        "user": user.json(),
+    }
+    if company:
+        resp["company"] = company.json()
+
+    return web.json_response(resp, status=HTTPStatus.OK)
+