+++ backend/database/__init__.py
@@ -3,20 +3,29 @@ from sqlalchemy.orm import relationship, Mapped
 
 from .models import Base
 
+from globalTypes import (
+    RegistrationType,
+    Role,
+    AuthStatus,
+    ConfirmStatusCode,
+    Remote,
+    SalaryType,
+    WorkingHours,
+    TravelRequired,
+    PositionStatus,
+)
+
 # models
-from .models.admin import Staff
-from .models.candidate import Candidate, CandidateExperience
-from .models.company import Company, CompanyManager
-from .models.user import User, ConfirmCode
-from globalTypes import AuthStatus, ConfirmStatusCode, RegistrationType, Role
+from .models import (
+    Staff,
+    Candidate, CandidateExperience,
+    Company, CompanyManager,
+    User, ConfirmCode,
+    Position, PositionViews, PositionFeedback
+)
 
-User.candidate = relationship("Candidate", back_populates="user")
-Candidate.user: Mapped["User"] = relationship(
-    "User", back_populates="candidate")
 
-User.manager = relationship("CompanyManager", back_populates="user")
-CompanyManager.user: Mapped["User"] = relationship(
-    "User", back_populates="manager")
-
+# User.feedbacks = relationship("PositionFeedback", uselist=True, back_populates="user")
+# PositionFeedback.user = relationship("User", back_populates="feedbacks")
 
 Base.registry.configure()