+++ backend/database/repositories/confirm_code.py
@@ -7,11 +7,13 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 
-async def get_confirm_code(session: AsyncSession, id: str, code: str, status: ConfirmStatusCode) -> Optional[ConfirmCode]:
+async def get_confirm_code(
+    session: AsyncSession, id: str, code: str, status: ConfirmStatusCode
+) -> Optional[ConfirmCode]:
     query = select(ConfirmCode).where(
-        (ConfirmCode.id == id) &
-        (ConfirmCode.status == status) &
-        (ConfirmCode.code == code)
+        (ConfirmCode.id == id)
+        & (ConfirmCode.status == status)
+        & (ConfirmCode.code == code)
     )
     result = await session.execute(query)
     fetched = result.fetchone()