+++ backend/database/repositories/registration.py
@@ -3,7 +3,7 @@ from typing import Tuple, Optional
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from api.types import RegistrationPayload
+from api.routes.public.types import RegistrationPayload
 from database import (
     Company,
     CompanyManager,
@@ -22,7 +22,7 @@ async def create_company(
     user = User(
         name=payload.name,
         email=payload.email,
-        role=Role.COMPANY_MANAGER,
+        role=Role.COMPANY_ADMIN,
         status=AuthStatus.PENDING,
         password_hash=password.get_hashed_password(payload.password),
         codes=[