+++ client/projects/client/src/app/modules/auth/containers/register-company/register-company.component.ts
@@ -1,8 +1,9 @@
-import { Component, computed, signal } from '@angular/core';
+import { Component } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Store } from '@ngrx/store';
 import { AppState } from '../../../../store';
 import { ActionRegisterCompanyStart } from '../../../../store/account';
+import { confirmPasswordValidator } from '@ui';
 
 @Component({
     selector: 'auth-register',
@@ -15,6 +16,10 @@ export class RegisterCompanyComponent {
     ngOnInit() {
         this.form = this.fb.group({
             name: ['', [Validators.required]],
+            email: ['', [Validators.required, Validators.email]],
+            password: ['', Validators.required],
+            password_confirm: ['', Validators.required, confirmPasswordValidator('password')],
+            company_name: ['', Validators.required],
         });
     }
 
@@ -23,6 +28,11 @@ export class RegisterCompanyComponent {
         this.store.dispatch(ActionRegisterCompanyStart({
             payload: {
                 name: this.form.controls['name'].value,
+                email: this.form.controls['email'].value,
+                password: this.form.controls['password'].value,
+                password_confirm: this.form.controls['password_confirm'].value,
+                type: 'company',
+                companyName: this.form.controls['company_name'].value,
             }
         }))
     }