+++ client/projects/client/src/app/store/account/account.actions.ts
@@ -1,6 +1,5 @@
 import { createAction, props } from "@ngrx/store";
 import { RegistrationPayload } from "../../models";
-import { RegistrationCompanyPayload } from "../../models/company";
 
 export const ActionRegisterAccountStart = createAction(
     "[account] register account start",
@@ -9,5 +8,5 @@ export const ActionRegisterAccountStart = createAction(
 
 export const ActionRegisterCompanyStart = createAction(
     "[account] register company start",
-    props<{ payload: RegistrationCompanyPayload }>(),
+    props<{ payload: RegistrationPayload }>(),
 )